+++ frontend/src/components/Modals/Modals.tsx
@@ -1,5 +1,4 @@
 import { WorkoutEditorModal } from './WorkoutEditorModal';
-import * as React from 'react';
 import { LogModal } from './LogModal';
 import { LoginModal } from './LoginModal';
 import { ProfileModal } from './ProfileModal';
@@ -8,10 +7,10 @@ import { GroupSessionModal } from './GroupSessionModal';
 export const Modals = () => {
   return (
     <>
+      <GroupSessionModal />
       <LogModal />
       <LoginModal />
       <ProfileModal />
-      <GroupSessionModal />
       <WorkoutEditorModal />
     </>
   );
