+++ frontend/src/components/TopBar.tsx
@@ -1,17 +1,16 @@
-import * as React from 'react';
 import { Center, Grid, Stack, Text } from '@chakra-ui/layout';
 import { hrColor, powerColor } from '../colors';
 import { useSmartTrainer } from '../context/SmartTrainerContext';
 import { useHeartRateMonitor } from '../context/HeartRateContext';
 import { useColorModeValue } from '@chakra-ui/color-mode';
 import theme from '../theme';
-import { secondsToHoursMinutesAndSecondsString } from '../utils';
 import { useData } from '../context/DataContext';
 import {
   getRemainingTime,
   getTargetPower,
   useActiveWorkout,
 } from '../context/ActiveWorkoutContext';
+import { secondsToHoursMinutesAndSecondsString } from '../utils/time';
 
 export const TopBar = () => {
   const { power, cadence } = useSmartTrainer();
