+++ frontend/src/utils/index.ts
@@ -1,0 +1,6 @@
+import * as general from './general';
+import * as gps from './gps';
+import * as time from './time';
+import * as zones from './zones';
+
+export default { general, gps, time, zones };
